Todos os desafios do 005 á 013 corrigidos e atualizados

+++ PythonDesafios/ex007.py
@@ -1,4 +1,4 @@
-n1 = int(input("Digite sua primeira nota: "))
-n2 = int(input("Digite sua segunda  nota: "))
+n1 = float(input("Digite sua primeira nota: "))
+n2 = float(input("Digite sua segunda  nota: "))
 print("A média das suas notas é: {}".format((n1+n2)/2))
